pyproject.toml
[project]
name = "insurance-rag"
version = "0.2.0"
requires-python = ">=3.10"
dependencies = [
    # Core RAG dependencies (existing)
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.0",
    "qdrant-client>=1.7.0",
    "pandas>=2.0.0",
    "python-dotenv>=1.0.0",
    
    # New Agentic dependencies
    "langgraph>=0.2.0",           # For building agents
    "langchain-tavily>=0.2.0",    # Official Tavily integration
    "ragas==0.2.10",               # For RAG evaluation with testset generation
    "langsmith>=0.1.0",           # For monitoring
    "Pillow>=10.0.0",             # PIL for RAGAS
    "datasets>=2.14.0",           # For dataset handling
    
    # Query enhancement
    "langchain-cohere>=0.1.0",    # For reranking (optional)
    "rank-bm25>=0.2.0",           # For keyword search
    
    # Helpful additions
    "rich>=13.0.0",               # Pretty console output
    "typer>=0.9.0",               # CLI interface
    
    # UI Dependencies
    "streamlit>=1.28.0",          # Web UI framework
    "plotly>=5.17.0",             # Interactive charts
    "streamlit-plotly-events>=0.0.6",  # Enhanced plotly integration
    "gradio>=4.0.0",              # Alternative UI framework
    "flask>=2.3.0",               # Web framework for custom UI
]

[tool.uv]
dev-dependencies = [
    "ipykernel>=6.0.0",          # For Jupyter
    "pytest>=7.0.0",             # For testing
]